chore: move 'Point' under its own module.
